Structure refactored
